requirements.txt
pandas>=2.0.0
requests>=2.28.0
openpyxl>=3.1.0
dbt-core >=1.5.0
dbt-snowflake >=1.5.0
snowflake-connector-python >=3.1.0
apache-airflow-providers-snowflake